move new endpoint to its own file
solves  https://github.com/fractal-analytics-platform/fractal-server/pull/1607#issuecomment-2210263749
